Починен основной файл digger после экспериментов с заставкой и звуком.

+++ sound.h
@@ -15,4 +15,4 @@ void sound(uint16_t period, uint16_t durance);
  * @param period - период звука (1 / частота)
  * @param durance - длительность звука
  */
-void sound_vibrato(uint16_t period, uint8_t durance);
+void sound_vibrato(uint16_t period, uint16_t durance);
